+++ CHANGELOG.md
@@ -19,6 +19,6 @@
 - Allow `trailingAnimation` supports any `Animatable` type and not just `Tween<double>`.
 - Added `separatorBuilder` property to customize the separator between the tiles
 - Added `ExpansionMode` property to specify the expansion mode of the `ExpansionTileList`
-- Added `ExpansionMode.single` to expand only one tile at a time.
-- Added `ExpansionMode.multiple` to expand multiple tiles at a time.
+- Added `ExpansionMode.single` named constructor to expand only one tile at a time.
+- Added `ExpansionMode.multiple` named constructor to expand multiple tiles at a time.
 - Added a new Widget `ExpansionTileItem` to customize the tile appearance for better flexibility when compared to `ExpansionTile`